+++ react/typings/vtex.recommendation-context.d.ts
@@ -1,0 +1,3 @@
+declare module 'vtex.recommendation-context/RecommendationContext' {
+  export const useRecommendation
+}