+++ react/typings/vtex.render-runtime.d.ts
@@ -2,4 +2,5 @@ declare module 'vtex.render-runtime' {
   export const ExtensionPoint
   export const useTreePath
   export const Loading
+  export const useRuntime
 }