+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- View and Product click events.
+
 ## [0.5.1] - 2020-10-23
 
 ### Fixed